requirements.txt
langchain
langchain_core
langchain_groq
python-dotenv
requests